register and edit profile

+++ routes/web.php
@@ -3,6 +3,8 @@
 use App\Http\Controllers\UkmController;
 use App\Http\Controllers\UkmRegistrationController;
 use App\Http\Controllers\ArticleController;
+use App\Http\Controllers\LoginController;
+use App\Http\Controllers\RegisterController;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -26,26 +28,16 @@ Route::get('/csrf-token', function () {
 
 Auth::routes();
 
-/* User */
-Route::middleware(['auth', 'user-access:user'])->group(function () {
-    Route::get('/home', [HomeController::class, 'index']);
-});
-
-/* Admin */
-Route::middleware(['auth', 'user-access:admin'])->group(function () {
-    Route::get('/dashboard', [HomeController::class, 'dashboard']);
-});
-
 Route::prefix('ukms')->group(function(){
-    Route::get('/register', [UkmRegistrationController::class, 'register']);
+    Route::get('/register', [UkmRegistrationController::class, 'register'])->middleware(['auth', 'user-access:user']);
     Route::prefix('registrations')->group(function(){
         Route::get('/', [UkmRegistrationController::class, 'read']);
-        Route::get('/ukm/{ukm_id}', [UkmRegistrationController::class, 'readUkm']);
-        Route::get('/status/{ukm_id}', [UkmRegistrationController::class, 'checkStatus']);
-        Route::post('/status/{ukm_id}', [UkmRegistrationController::class, 'changeStatus']);
-        Route::post('/column/add', [UkmRegistrationController::class, 'addColumn']);
-        Route::post('/column/delete', [UkmRegistrationController::class, 'deleteColumn']);
-        Route::get('/description/{ukm_id}', [UkmRegistrationController::class, 'getRegistDescription']);
+        Route::get('/ukm/{ukm_id}', [UkmRegistrationController::class, 'readUkm'])->middleware(['auth', 'user-access:admin']);
+        Route::get('/status/{ukm_id}', [UkmRegistrationController::class, 'checkStatus'])->middleware(['auth', 'user-access:admin']);
+        Route::post('/status/{ukm_id}', [UkmRegistrationController::class, 'changeStatus'])->middleware(['auth', 'user-access:admin']);
+        Route::post('/column/add', [UkmRegistrationController::class, 'addColumn'])->middleware(['auth', 'user-access:admin']);
+        Route::post('/column/delete', [UkmRegistrationController::class, 'deleteColumn'])->middleware(['auth', 'user-access:admin']);
+        Route::get('/description/{ukm_id}', [UkmRegistrationController::class, 'getRegistDescription'])->middleware(['auth', 'user-access:admin']);
         Route::get('/{id}', [UkmRegistrationController::class, 'readId']);
     });
 
@@ -60,13 +52,31 @@ Route::prefix('ukms')->group(function(){
 
 Route::prefix('articles')->group(function(){
     Route::get('/', [ArticleController::class, 'read']);
-    Route::post('/', [ArticleController::class, 'create']);
+    Route::post('/', [ArticleController::class, 'create'])->middleware(['auth', 'user-access:admin']);
     Route::get('/ukm/{ukm_id}', [ArticleController::class, 'readUkm']);
     Route::get('/category/{category}', [ArticleController::class, 'readCategory']);
     Route::get('/search/{key}', [ArticleController::class, 'search']);
-    Route::post('/edit/{id}', [ArticleController::class, 'update']);
+    Route::post('/edit/{id}', [ArticleController::class, 'update'])->middleware(['auth', 'user-access:admin']);
     Route::get('/{id}', [ArticleController::class, 'readId']);
-    Route::delete('/{id}', [ArticleController::class, 'delete']);
+    Route::delete('/{id}', [ArticleController::class, 'delete'])->middleware(['auth', 'user-access:admin']);
 });
 
 Auth::routes(['verify' => true]);
+
+Route::prefix('register')->group(function(){
+    Route::get('/', [RegisterController::class, 'index']);
+    Route::post('/', [RegisterController::class, 'store']);
+});
+
+Route::prefix('login')->group(function(){
+    Route::get('/', [LoginController::class, 'login'])->middleware('guest');
+    Route::post('/', [LoginController::class, 'user']);
+    Route::post('/admin', [LoginController::class, 'admin']);
+});
+
+Route::post('/logout', [LoginController::class, 'logout']);
+
+Route::prefix('profiles')->group(function(){
+    Route::get('/', [ArticleController::class, 'read'])->middleware(['auth', 'user-access:user']);
+    Route::post('/', [ArticleController::class, 'update']);
+});
